% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konwersja_tempr.R
\name{konwersja_temp}
\alias{konwersja_temp}
\title{Konwersja temperatur}
\usage{
konwersja_temp(temperatura_f)
}
\arguments{
\item{Ta}{funkcja przelicza stopnie C na F}
}
\value{
wektor o takiej samiej dlugosci jak temperatura_f
}
\description{
Konwersja temperatur
}
\examples{
konwersja_temp(50)
}
